+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/models/QueryUsersRequest.kt
@@ -16,13 +16,15 @@
 
 package io.getstream.chat.android.client.api.models
 
+import io.getstream.chat.android.client.api.models.querysort.QuerySortByField
+import io.getstream.chat.android.client.api.models.querysort.QuerySorter
 import io.getstream.chat.android.client.models.User
 
 public data class QueryUsersRequest @JvmOverloads constructor(
     var filter: FilterObject,
     val offset: Int,
     val limit: Int,
-    var querySort: QuerySort<User> = QuerySort(),
+    var querySort: QuerySorter<User> = QuerySortByField(),
     var presence: Boolean = false
 ) {
     val sort: List<Map<String, Any>> = querySort.toDto()