+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/ScrollButtonViewStyle.kt
@@ -25,6 +25,7 @@ import androidx.annotation.StyleableRes
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.TransformStyle
 import io.getstream.chat.android.ui.common.extensions.internal.getColorCompat
+import io.getstream.chat.android.ui.common.extensions.internal.getColorOrNull
 import io.getstream.chat.android.ui.common.extensions.internal.getDimension
 import io.getstream.chat.android.ui.common.style.TextStyle
 import io.getstream.chat.android.ui.message.list.internal.ScrollButtonView
@@ -36,29 +37,49 @@ import io.getstream.chat.android.ui.message.list.internal.ScrollButtonView
  * @property scrollButtonUnreadEnabled Enables/disables unread label. Default value is enabled.
  * @property scrollButtonColor [ScrollButtonView] background color. Default value is [R.color.stream_ui_white].
  * @property scrollButtonRippleColor [ScrollButtonView] ripple color. Default value is [R.color.stream_ui_white_smoke].
- * @property scrollButtonBadgeColor Unread label background color. Default value is [R.color.stream_ui_accent_blue].
+ * @property scrollButtonBadgeColor Unread count label background color. Default value is [R.color.stream_ui_accent_blue].
  * @property scrollButtonIcon [ScrollButtonView] icon. Default value is [R.drawable.stream_ui_ic_down].
- * @property scrollButtonBadgeTextStyle Appearance for unread label.
+ * @property scrollButtonBadgeTextStyle Text appearance of the unread count label.
+ * @property scrollButtonBadgeIcon The icon used for the unread count label. Default value is [R.drawable.stream_ui_shape_scroll_button_badge].
+ * @property scrollButtonBadgeGravity The gravity of the unread count label. Default value is center top.
+ * @property scrollButtonBadgeElevation The elevation used for the unread count label.
+ * Default value is [MessageListViewStyle.DEFAULT_SCROLL_BUTTON_BADGE_ELEVATION].
+ * @property scrollButtonInternalMargin The internal margins for the scroll button.
+ * This affects the distance between the scroll button and the unread count label.
  */
 public data class ScrollButtonViewStyle(
     public val scrollButtonEnabled: Boolean,
     public val scrollButtonUnreadEnabled: Boolean,
     @ColorInt public val scrollButtonColor: Int,
     @ColorInt public val scrollButtonRippleColor: Int,
-    @ColorInt public val scrollButtonBadgeColor: Int,
+    @ColorInt public val scrollButtonBadgeColor: Int?,
     public val scrollButtonElevation: Float,
     public val scrollButtonIcon: Drawable?,
     public val scrollButtonBadgeTextStyle: TextStyle,
+    public val scrollButtonBadgeIcon: Drawable?,
+    public val scrollButtonBadgeGravity: Int,
+    public val scrollButtonBadgeElevation: Float,
+    public val scrollButtonInternalMargin: Int,
 ) {
 
     internal class Builder(private val context: Context, private val attrs: TypedArray) {
         private var scrollButtonEnabled: Boolean = false
         private var scrollButtonUnreadEnabled: Boolean = false
-        @ColorInt private var scrollButtonColor: Int = 0
-        @ColorInt private var scrollButtonRippleColor: Int = 0
-        @ColorInt private var scrollButtonBadgeColor: Int = 0
+
+        @ColorInt
+        private var scrollButtonColor: Int = 0
+
+        @ColorInt
+        private var scrollButtonRippleColor: Int = 0
+
+        @ColorInt
+        private var scrollButtonBadgeColor: Int? = 0
         private var scrollButtonIcon: Drawable? = null
         private var scrollButtonElevation: Float = 0F
+        private var scrollButtonBadgeGravity: Int = 0
+        private var scrollButtonBadgeIcon: Drawable? = null
+        private var scrollButtonBadgeElevation: Float = 0f
+        private var scrollButtonInternalMargin: Int = 0
 
         fun scrollButtonEnabled(
             @StyleableRes scrollButtonEnabledStyleableId: Int,
@@ -90,9 +111,8 @@ public data class ScrollButtonViewStyle(
 
         fun scrollButtonBadgeColor(
             @StyleableRes scrollButtonBadgeColorStyleableId: Int,
-            @ColorInt defaultColor: Int,
         ) = apply {
-            scrollButtonBadgeColor = attrs.getColor(scrollButtonBadgeColorStyleableId, defaultColor)
+            scrollButtonBadgeColor = attrs.getColorOrNull(scrollButtonBadgeColorStyleableId)
         }
 
         fun scrollButtonIcon(
@@ -107,6 +127,34 @@ public data class ScrollButtonViewStyle(
             defaultElevation: Float,
         ) = apply {
             this.scrollButtonElevation = attrs.getDimension(scrollButtonElevation, defaultElevation)
+        }
+
+        fun scrollButtonBadgeGravity(
+            @StyleableRes scrollButtonBadgeGravity: Int,
+            defaultGravity: Int,
+        ): Builder = apply {
+            this.scrollButtonBadgeGravity = attrs.getInt(scrollButtonBadgeGravity, defaultGravity)
+        }
+
+        fun scrollButtonBadgeIcon(
+            @StyleableRes scrollButtonBadgeIcon: Int,
+            defaultIcon: Drawable?,
+        ) = apply {
+            this.scrollButtonBadgeIcon = attrs.getDrawable(scrollButtonBadgeIcon) ?: defaultIcon
+        }
+
+        fun scrollButtonBadgeElevation(
+            @StyleableRes scrollButtonBadgeElevation: Int,
+            defaultElevation: Float,
+        ) = apply {
+            this.scrollButtonBadgeElevation = attrs.getDimension(scrollButtonBadgeElevation, defaultElevation)
+        }
+
+        fun scrollButtonBadgeInternalMargin(
+            @StyleableRes scrollButtonInternalMargin: Int,
+            defaultMargin: Int,
+        ) = apply {
+            this.scrollButtonInternalMargin = attrs.getDimensionPixelSize(scrollButtonInternalMargin, defaultMargin)
         }
 
         fun build(): ScrollButtonViewStyle {
@@ -134,7 +182,11 @@ public data class ScrollButtonViewStyle(
                 scrollButtonBadgeColor = scrollButtonBadgeColor,
                 scrollButtonIcon = scrollButtonIcon,
                 scrollButtonBadgeTextStyle = scrollButtonBadgeTextStyle,
-                scrollButtonElevation = scrollButtonElevation
+                scrollButtonElevation = scrollButtonElevation,
+                scrollButtonBadgeIcon = scrollButtonBadgeIcon,
+                scrollButtonBadgeGravity = scrollButtonBadgeGravity,
+                scrollButtonBadgeElevation = scrollButtonBadgeElevation,
+                scrollButtonInternalMargin = scrollButtonInternalMargin
             ).let(TransformStyle.scrollButtonStyleTransformer::transform)
         }
     }